gcc-4.2.4: fix libstdc++_pic for Meta
It is not clear why 300-libstdc++-pic.patch pre-applied to gcc-4.2.4
seems to have come from some older version of gcc (as early as 4.0.3).
This older patch incorrectly refers to object files to be included in
the library libstdc++_pic.a as *.o, while their location seems to be in
.libs/*.o (see the contents at e.g.
https://github.com/ZigFisher/Midge/blob/master/toolchain/gcc/4.0.3/300-libstdc%2B%2B-pic.patch).
The same patch for gcc 4.1.0 refers to the files as .libs/*.o (see
https://github.com/ZigFisher/Midge/blob/master/toolchain/gcc/4.1.0/300-libstdc%2B%2B-pic.patch).
This patch corrects rules in Makefile.in.

This is based on commit 24c5dbcef21d (gcc: fix libstdc++_pic for AVR32).

Signed-off-by: James Hogan <james.hogan@imgtec.com>